Prepend QTGRAPHS_ to include guards
Datavis and Graphs have many duplicated files with same include guards.
Because QtGraphs has inherited many of its classes
from QtDataVisualization, the modules have many classes with the same
name. Therefore, mixing QtDataVisualization and QtGraphs headers in
the same TU violates ODR. This patch includes the check for that and
fails compilation with an error message.

Pick-to: 6.9
Change-Id: I245c71c3fc5661ce0c33d5cd751ffb2a04754ad1
Reviewed-by: Marc Mutz <marc.mutz@qt.io>