Expose qabstractseries series as anonymous
If a type is not exposed to qml, having revision macro in the type's
signals causes problems in subtypes that are exposed to qml.
This patch exposes qabstractseries as an anonymous type.
In addition fix missing " problems in qmltestbed.
Amends 28df8f7f3764cee39a3fa8ee294d38b07e27ad3d

Change-Id: I21573713d5096df961fabd4ee9c1e8d8932c49bb
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>